temporal_knowledge_base/pyproject.toml
[project]
name = "chronos"
version = "0.1.0"
description = "Temporal knowledge base for leakage-free leader decision prediction — time-slidable, multi-model, autonomously constructed"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "langchain-google-genai>=4.0",
    "langchain-community>=0.3",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "pgvector>=0.3",
    "newspaper3k>=0.2",
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
    "python-dateutil>=2.9",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "numpy>=1.26",
    "aiofiles>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
